Misc manual react-doctor resolutions
