+++ src/common/components/Header/types.ts
@@ -1,6 +1,10 @@
+export type SubPageInfo = {
+  pageName: string;
+  link: string;
+};
+
 export type PageInfo = {
   pageName: string;
   link: string;
-  dropdownID?: number;
-  subPages?: { pageName: string; link: string }[];
+  subPages?: SubPageInfo[];
 };
