+++ src/common/hooks/ui/index.ts
@@ -4,4 +4,5 @@ export * from "./useHeaderHeight";
 export * from "./useGrowOnScroll";
 export * from "./usePlotDownload";
 export * from "./useScrollReset";
+export * from "./useScrollToSearch";
 export * from "./useTablePlotSync";
